pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Bahadır Araz", email = "bahadiraraz@protonmail.com"},
]
dependencies = [
    "fastapi>=0.105.0",
    "pymongo>=4.6.1",
    "pydantic>=2.5.2",
    "web3>=6.12.0",
    "python-dotenv>=1.0.0",
    "pinatapy-vourhey>=0.1.9",
    "loguru>=0.7.2",
    "qrcode[pil]>=7.4.2",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

